+++ uni_modules/vk-unicloud/changelog.md
@@ -1,3 +1,8 @@
+## 2.8.7（2022-05-14）
+* 1、【新增】`vk.pubfn.getFileType(url)` 根据文件url获取文件类型
+* 2、【新增】`vk.pubfn.getFileSuffix(url)` 根据文件url获取文件后缀名
+* 3、【新增】`vk.pubfn.checkFileSuffix(url,["png", "jpg", "jpeg", "gif", "bmp", "svg"])` 根据文件url判断是否满足指定后缀名中的任意1个
+* 完整框架项目地址：`https://ext.dcloud.net.cn/plugin?id=2204`[点击查看](https://ext.dcloud.net.cn/plugin?id=2204)
 ## 2.8.6（2022-05-13）
 * 1、【优化】一些细节
 * 完整框架项目地址：`https://ext.dcloud.net.cn/plugin?id=2204`[点击查看](https://ext.dcloud.net.cn/plugin?id=2204)